Get token on account change
