+++ src/containers/Publish/style.scss
@@ -8,6 +8,11 @@
   align-items: center;
 }
 
+div.tokenControl {
+  margin-bottom: 10px;
+  min-width: 30px;
+}
+
 .menuItem {
   display: flex;
 }
